+++ src/main/java/dinkplugin/domain/LootCriteria.java
@@ -1,0 +1,8 @@
+package dinkplugin.domain;
+
+public enum LootCriteria {
+    ALLOWLIST,
+    DENYLIST,
+    VALUE,
+    RARITY
+}