+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Minor: Add clan title to chat notification metadata. (#635)
 - Minor: Allow hiding just split private chat from screenshots. (#630)
 - Dev: Remove deprecated `getCachedNPCs`/`getCachedPlayers` calls. (#632)